Initial commit
this program loads and displays a bitmap onto an arduino compatible TFT
LCD screen v1.7 includes lifestages for the tamagotchi (egg, hatched,
dead) and supports a temperature sensor to mimick good conditions to
hatch faster, number of lifes, feeding and dying. Also included is a
push button to feed the tamagotchi to prevent it from dying. It is
writen in C (Arduino)